Update configs and builds for ARP Intel
Updated docker configs so that it will be easier to use both in pure docker install
and in mixed docker/developer install, where infra runs on docker, while the
microservices run locally. For this, eg. mysql must be accessible both from
docker and from the outside. Also needed to update some libs, images to work with both
intel and arm (M1 mac) docker.

- Mysql users are not limited to localhost, but we use '%' so that we can access
  the database from anywhere (from inside and outside docker)
- Upgrade gosu to 1.14
- Use neo4j:3.5.6 instead of 3.5.3 so that it will be usable on both intel and arm
- Some hacking around openview's javascript location
- Added missing env vars to cedar-microservices/docker-compose.yml