Add devinstall.sh
devinstall.sh can be used to setup a complete CEDAR dev env running the
infra services in Docker, while running the microservices natively hence
allowing it to be locally developed.

devinstall.sh should be run once and should build the docker containers,
run them, then build the microservices, run them and configure their
initial users and contents.

If all goes well and the frontend is run using gulp the local CEDAR
should be available at https://cedar.metadatacenter.orgx

This is the first iteration of the script and needs more options to be
more usable and versatile.